+++ protocols/request-response/CHANGELOG.md
@@ -3,6 +3,9 @@
 - fix: public cbor/json codec module
   See [PR 5830](https://github.com/libp2p/rust-libp2p/pull/5830).
 
+- feat: add `Behaviour::send_request_with_addresses()`
+  See [PR 5938](https://github.com/libp2p/rust-libp2p/issues/5938).
+
 ## 0.28.0
 
 - Deprecate `void` crate.