+++ protocols/request-response/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.29.0
+
+- Use `tokio` for libp2p-tcp as `async-std` support has been removed.
+  See [PR 5955](https://github.com/libp2p/rust-libp2p/pull/5955).
+
+
 ## 0.28.1
 
 - fix: public cbor/json codec module